group module

+++ internal/domain/message/message_repository.go
@@ -2,5 +2,5 @@ package message
 
 type MessageRepository interface {
 	Save(msg *Message) (*Message, error)
-	GetMessagesForChat(chatId int64, since int) ([]Message, error)
+	GetMessagesForChat(chatId int64, since int64) ([]Message, error)
 }
